// src/frontend/src/backend.d.ts
export type Time = bigint;
export interface SmartRuleLog {
    ruleName: string;
    triggerCondition: string;
    timestamp: Time;
    outcome: boolean;
    actionTaken: string;
}
export interface SmartRule {
    ruleName: string;
    action: string;
    active: boolean;
    createdBy: string;
    deviceId: string;
    users: Array<string>;
    schedule: string;
    triggers: Array<string>;
}
export interface DeviceStatus {
    batteryLevel: bigint;
    tamperDetected: boolean;
    wifiSignalStrength: bigint;
    isLocked: boolean;
    emergencyMode: boolean;
    firmwareVersion: string;
    lastSync: bigint;
    currentUsers: bigint;
}
export interface Subtask {
    id: bigint;
    text: string;
    completed: boolean;
}
export interface DeviceInfo {
    status: DeviceStatus;
    isConnected: boolean;
    deviceId: string;
    deviceName: string;
    lastSync: Time;
}